+++ VS/Graphics/VS_Scene.cpp
@@ -32,7 +32,7 @@ public:
 
 	void				Init()
 	{
-		SetFOV( vsScreen::Instance()->GetHeight() );
+		SetFOV( (float)vsScreen::Instance()->GetHeight() );
 		SetPosition( vsVector2D( vsScreen::Instance()->GetWidth() * 0.5f,
 					vsScreen::Instance()->GetHeight() * 0.5f ) );
 	}
